AWS-RS-HW-5 Refactored
