--- cdk-ts/utils/response.ts
@@ -1,15 +1,0 @@
-export const setResponse = (
-    statusCode: number = 200,
-    body = {},
-) => {
-    return {
-        statusCode,
-        headers: {
-            "Access-Control-Allow-Origin": "*",
-            "Access-Control-Allow-Headers": "*",
-            "Access-Control-Allow-Methods": "*",
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify(body),
-    };
-};
